requirements.txt
bs4
flask
numpy
requests
pandas
nltk
sklearn
slackclient
slackeventsapi
python-dotenv